Refactor: Extract GitHub Integration component into its own module
